Added HasRack mixin and updated data model.
